requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.0
PyAutoGUI~=0.9.54